requirements.txt
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
scipy==1.11.4

# API and web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Data storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging and monitoring
loguru==0.7.2

# Jupyter for analysis

matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
click==8.1.7